src/pyproject.toml
[tool.poetry]
name = "medical-image-captioning"
version = "0.1.0"
description = ""
authors = ["Sally"]
readme = "README.md"
packages = [{include = "medical_image_captioning"}]

[tool.poetry.dependencies]
python = "3.8.16"
python-dotenv = "^0.21.0"
transformers = "^4.25.1"
torch = "^1.13.1"
torchvision = "^0.14.1"
hydra-core = "^1.3.1"
pytorch-lightning = "^1.8.6"
torchmetrics = "^0.11.0"
pillow = "^9.3.0"
rich = "^12.6.0"
hydra-colorlog = "^1.2.0"
pandas = "^1.5.2"
sklearn = "^0.0.post1"
scikit-learn = "^1.2.0"
scipy = "^1.9.3"
matplotlib = "^3.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
